added rpc and DB actions

+++ src/ru/rs/testcase/phones/server/data/ContactsDaoImpl.java
@@ -17,7 +17,7 @@ public class ContactsDaoImpl implements ContactsDao {
 
     @Override
     public Contact createContact(Contact contact) {
-
+    
         try (Connection connect = helper.getConnection()) {
             connect.setAutoCommit(false);
             PreparedStatement st =
@@ -38,21 +38,26 @@ public class ContactsDaoImpl implements ContactsDao {
 
     @Override
     public List<Contact> findAllContacts() throws ContactsException {
-
+    
         List<Contact> result = new ArrayList<>();
         try (Connection c = helper.getConnection()) {
-            Statement st = c.createStatement();
-            ResultSet rs = st.executeQuery("SELECT NAME,PHONE FROM PHONES");
-            while (rs.next()) {
-                Contact contact = new Contact();
-
-                contact.setName(rs.getString("NAME"));
-                contact.setPhone(rs.getString("PHONE"));
-                result.add(contact);
+            if (c != null) {
+                Statement st = c.createStatement();
+                ResultSet rs = st.executeQuery("SELECT NAME,PHONE FROM PHONES");
+                while (rs.next()) {
+                    Contact contact = new Contact();
+                    
+                    contact.setName(rs.getString("NAME"));
+                    contact.setPhone(rs.getString("PHONE"));
+                    result.add(contact);
+                }
+                c.commit();
+                rs.close();
+                st.close();
+            } else {
+                throw new ContactsException(
+                        "Не удалось подключиться к Базе данных");
             }
-            c.commit();
-            rs.close();
-            st.close();
         } catch (SQLException e) {
             throw new ContactsException("Unable to get all contacts");
         }
@@ -63,23 +68,28 @@ public class ContactsDaoImpl implements ContactsDao {
     public List<Contact> findBySubstring(String subString)
             throws ContactsException
     {
-
+    
         List<Contact> result = new ArrayList<>();
         try (Connection connection = helper.getConnection()) {
-            PreparedStatement ps =
-                    connection
-                    .prepareStatement("SELECT NAME,PHONE FROM PHONES WHERE NAME LIKE ?");
-            ps.setString(1, subString.concat("%"));
-            ResultSet set = ps.executeQuery();
-            while (set.next()) {
-                Contact contact = new Contact();
-                contact.setName(set.getString("NAME"));
-                contact.setPhone(set.getString("PHONE"));
-                result.add(contact);
+            if (connection != null) {
+                PreparedStatement ps =
+                        connection
+                                .prepareStatement("SELECT NAME,PHONE FROM PHONES WHERE NAME LIKE ?");
+                ps.setString(1, subString.concat("%"));
+                ResultSet set = ps.executeQuery();
+                while (set.next()) {
+                    Contact contact = new Contact();
+                    contact.setName(set.getString("NAME"));
+                    contact.setPhone(set.getString("PHONE"));
+                    result.add(contact);
+                }
+                connection.commit();
+                set.close();
+                ps.close();
+            } else {
+                throw new ContactsException(
+                        "Не удалось подключиться к Базе данных");
             }
-            connection.commit();
-            set.close();
-            ps.close();
         } catch (SQLException e) {
             throw new ContactsException("Unable to get filtered contacts");
         }
@@ -88,40 +98,50 @@ public class ContactsDaoImpl implements ContactsDao {
 
     @Override
     public void updateContact(Contact contact) throws ContactsException {
-
+    
         try (Connection connection = helper.getConnection()) {
-            PreparedStatement ps =
-                    connection
-                    .prepareStatement("UPDATE PHONES SET NAME=?,PHONE=? WHERE NAME=?");
-            ps.setString(1, contact.getName());
-            ps.setString(2, contact.getPhone());
-            ps.setString(3, contact.getName());
-            ps.executeUpdate();
-            connection.commit();
-            ps.close();
+            if (connection != null) {
+                PreparedStatement ps =
+                        connection
+                                .prepareStatement("UPDATE PHONES SET NAME=?,PHONE=? WHERE NAME=?");
+                ps.setString(1, contact.getName());
+                ps.setString(2, contact.getPhone());
+                ps.setString(3, contact.getName());
+                ps.executeUpdate();
+                connection.commit();
+                ps.close();
+            } else {
+                throw new ContactsException(
+                        "Не удалось подключиться к Базе данных");
+            }
         } catch (SQLException e) {
             throw new ContactsException("Unable to Edit contact");
         }
     }
-
+    
     @Override
     public void removeContact(Contact contact) throws ContactsException {
-
+    
         try (Connection connection = helper.getConnection()) {
-            PreparedStatement ps =
-                    connection
-                    .prepareStatement("DELETE FROM PHONES WHERE NAME=? AND PHONE=?");
-            ps.setString(1, contact.getName());
-            ps.setString(2, contact.getPhone());
-            ps.executeUpdate();
-            connection.commit();
-            ps.close();
+            if (connection != null) {
+                PreparedStatement ps =
+                        connection
+                                .prepareStatement("DELETE FROM PHONES WHERE NAME=? AND PHONE=?");
+                ps.setString(1, contact.getName());
+                ps.setString(2, contact.getPhone());
+                ps.executeUpdate();
+                connection.commit();
+                ps.close();
+            } else {
+                throw new ContactsException(
+                        "Не удалось подключиться к Базе данных");
+            }
         } catch (SQLException e) {
             throw new ContactsException("Unable to delete contact" + contact);
         }
         
     }
-
+    
     /**
      * Dangerous, only for tests
      *
@@ -131,7 +151,7 @@ public class ContactsDaoImpl implements ContactsDao {
     @Deprecated
     @Override
     public void removeAll() throws ContactsException {
-
+    
         try (Connection connection = helper.getConnection()) {
             PreparedStatement ps =
                     connection.prepareStatement("DELETE FROM PHONES");
@@ -142,5 +162,5 @@ public class ContactsDaoImpl implements ContactsDao {
             throw new ContactsException("Unable to remove all contacts");
         }
     }
-
+    
 }
